Config explaination change in tesseract_eng.yml (#10)
* Config explaination change in tesseract_eng.yml

This is a huge change for tesseract_eng.yml, this further explains each setting in a more organized fashion and in a more better way. I myself feel this is very useful for players that are new to setting up a server, and to others who already have a server that is running Tesseract. I have made this into a separate branch so the other languages can be set to this (if approved) by the help of translators and to see what can be changed.

* fix extraneous space in inventory setting area

* my mistake

* Typo lol